reflector/requirements.txt
# ARIA-SDR Reflector - Dependências Simplificadas
agno
openai
python-dotenv
numpy
requests
uvicorn
